機能を外部クラスに切り出し
